requirements.txt
flask
pandas
numpy
scikit-learn
joblib
python-whois
requests